blas & lapack: 1. removed unused methods in lapack_connector.h 2. moved blas wrappers to blas_connector.h 3. define a new class BlasConnectors to include blas wrappers 4. remove some unused 'do nothing' methods from xc_3.cpp
