+++ source/module_base/inverse_matrix.h
@@ -15,7 +15,7 @@ class Inverse_Matrix_Complex
 	~Inverse_Matrix_Complex();
 	
 	ModuleBase::ComplexMatrix A;
-	void using_zpotrf( const ModuleBase::ComplexMatrix &Sin);
+
 	void using_zheev(const ModuleBase::ComplexMatrix &in, ModuleBase::ComplexMatrix &out);
 	void init( const int &dim_in);
 
@@ -31,17 +31,6 @@ class Inverse_Matrix_Complex
 	ModuleBase::ComplexMatrix EA;
 };
 
-//not been used yet!
-class Inverse_Matrix_Real
-{
-	public:
-	
-	Inverse_Matrix_Real(){};
-	~Inverse_Matrix_Real(){};
-
-	int using_spotri(matrix &A, const int dim);
-
-};
 
 }
 #endif
