+++ source/src_io/chi0_hilbert.cpp
@@ -2461,82 +2461,6 @@ std::complex<double> Chi0_hilbert:: Cal_g(int iq)
     return g;
 }
 
-
-//---------------------------------------------
-//  matrix inverse( exist some bugs!)
-//---------------------------------------------
-
-/*void Chi0_hilbert:: C_inverse(int n, std::complex<double>** a)
-{
-         int info;
-
-         //std::complex<double> M[n][n],N[n][n];
-         std::complex<double> M[n*n],N[n*n];
-
-         for(int i=0; i<n; i++)
-         {
-             for(int j=0; j<n; j++)
-             {
-                 int index = i * n + j;
-                 M[index] = a[i][j];
-                 //M[i][j] = a[i][j];
-             }
-         }
-
-         std::cout<<"a in "<<std::endl;
-         for(int i=0; i<n; i++)
-         {
-             for(int j=0; j<n; j++)
-             {
-                 int index = i * n + j;
-                 //std::cout <<M[i][j]<<" ";
-                 std::cout<<M[index]<<" ";
-             }
-             std::cout <<std::endl;
-         }
-
-         int dim = n;
-         int nrhs = n;
-         int lda = n;
-         int ipiv = n;
-         int ldb = n;
-
-         //zgesv_(&dim, &nrhs, M[0], &lda, &ipiv, N[0], &ldb, &info);
-         //zgesv_(&dim, &nrhs, M, &lda, &ipiv, N, &ldb, &info);
-
-         zgetrf_(&dim, &dim, M, &lda, &ipiv, &info);
-
-         zgetri_(&dim, M, &lda, &ipiv, N, &dim, &info);
-
-
-         std::cout<<"a out "<<std::endl;
-         for(int i=0; i<n; i++)
-         {
-             for(int j=0; j<n; j++)
-             {
-                 int index = i * n + j;
-                 //std::cout <<N[i][j]<<" ";
-                 std::cout <<M[index]<<" ";
-             }
-             std::cout <<std::endl;
-         }
-
-         for(int i=0; i<n; i++)
-         {
-             for(int j=0; j<n; j++)
-             {
-                 int index = i * n + j;
-                 //a[i][j] = N[i][j];
-                 a[i][j] = M[index];
-             }
-         }
-
-         //delete[] M;
-         //delete[] N;
-
-         return;
-}*/
-
 //----------------------------------------------------
 // plot function
 //----------------------------------------------------
