+++ source/src_io/istate_charge.cpp
@@ -4,6 +4,7 @@
 #include "../module_base/global_variable.h"
 #include "../src_parallel/parallel_common.h"
 #include "../module_base/scalapack_connector.h"
+#include "../module_base/blas_connector.h"
 
 IState_Charge::IState_Charge(){}
 
@@ -210,7 +211,7 @@ void IState_Charge::idmatrix(const int &ib)
 	
 			for(int ir=0; ir!=wg_wfc.nr; ++ir)
 			{
-				LapackConnector::scal( wg_wfc.nc, wg_local[ir], wg_wfc.c+ir*wg_wfc.nc, 1 );
+				BlasConnector::scal( wg_wfc.nc, wg_local[ir], wg_wfc.c+ir*wg_wfc.nc, 1 );
 			}
 
 			// C++: dm(iw1,iw2) = wfc(ib,iw1).T * wg_wfc(ib,iw2)
