+++ source/src_io/wf_local.cpp
@@ -497,7 +497,7 @@ void WF_Local::distri_lowf_new(double **ctot, const int &is)
 			const int inc=1;
 			if(myid==src_rank)
 			{
-				LapackConnector::copy(GlobalC::ParaO.nloc, work, inc, GlobalC::LOC.wfc_dm_2d.wfc_gamma[is].c, inc);
+				BlasConnector::copy(GlobalC::ParaO.nloc, work, inc, GlobalC::LOC.wfc_dm_2d.wfc_gamma[is].c, inc);
 			}
 		}//loop ipcol
 	}//loop	iprow
@@ -569,7 +569,7 @@ void WF_Local::distri_lowf_complex_new(std::complex<double> **ctot, const int &i
 			const int inc=1;
 			if(myid==src_rank)
 			{
-				LapackConnector::copy(GlobalC::ParaO.nloc, work, inc, GlobalC::LOC.wfc_dm_2d.wfc_k[ik].c, inc);
+				BlasConnector::copy(GlobalC::ParaO.nloc, work, inc, GlobalC::LOC.wfc_dm_2d.wfc_k[ik].c, inc);
 			}
 		}//loop ipcol
 	}//loop	iprow
