+++ source/src_lcao/LCAO_evolve.cpp
@@ -81,16 +81,6 @@ void Evolve_LCAO_Matrix::using_LAPACK_complex(const int &ik, std::complex<double
 */
 				
 	int INFO;
-/*	LapackConnector::zpotrf('U',GlobalV::NLOCAL,Stmp,GlobalV::NLOCAL,&INFO);
-	for (int i=0;i<GlobalV::NLOCAL; i++)
-	{
-		for (int j=0; j<GlobalV::NLOCAL; j++)
-		{
-			Stmp(i,j)=Stmp(j,i);
-		}
-	}
-	LapackConnector::zpotri('U',GlobalV::NLOCAL,Stmp,GlobalV::NLOCAL,&INFO);
-*/
 
         int LWORK=3*GlobalV::NLOCAL-1; //tmp
         std::complex<double> * WORK = new std::complex<double>[LWORK];
@@ -285,181 +275,6 @@ void Evolve_LCAO_Matrix::using_LAPACK_complex(const int &ik, std::complex<double
 
 	return;
 }
-
-/*
-#include <cmath>
-#include "mpi.h"
-extern "C"
-{
-    #include "/export/soft/intel2015/composer_xe_2015.1.133/mkl/include/mkl_pblas.h"
-    #include "/export/soft/intel2015/composer_xe_2015.1.133/mkl/include/mkl_blacs.h"
-    #include "/export/soft/intel2015/composer_xe_2015.1.133/mkl/include/mkl_scalapack.h"
-    #include "/export/soft/intel2015/composer_xe_2015.1.133/mkl/include/mkl_blas.h"
-}
-#include "../module_base/blas_connector.h"
-
-int Evolve_LCAO_Matrix::using_ScaLAPACK_complex(const int &ik, std::complex<double>** c, std::complex<double>** c_init)const
-{
-	ModuleBase::TITLE("Evolve_LCAO_Matrix", "using_scalapack_complex");
-	ModuleBase::ComplexMatrix Htmp(GlobalV::NLOCAL,GlobalV::NLOCAL);
-	ModuleBase::ComplexMatrix Stmp(GlobalV::NLOCAL,GlobalV::NLOCAL);
-	for(int i=0; i<GlobalV::NLOCAL; i++)
-        {
-        		for(int j=0; j<GlobalV::NLOCAL; j++)
-                	{
-                		Htmp(i,j) = GlobalC::LM.Hloc2[i*GlobalV::NLOCAL+j];
-                           	Stmp(i,j) = GlobalC::LM.Sloc2[i*GlobalV::NLOCAL+j];
-                	}
-        }
-
-        for(int i=0; i<GlobalV::NLOCAL; i++)
-        {
-                for(int j=0; j<GlobalV::NLOCAL; j++)
-                {
-                        std::cout << Stmp(i,j) <<"\t";
-                }
-                std::cout <<std::endl;
-        }
-	std::cout <<std::endl;
-
-	std::cout << "11111" << std::endl;
-
-	int nFull, nblk;
-	int myid, nprocs, myprow, nprows, mypcol, npcols;
-	int my_blacs_ctxt;
-	int mpierr, info;
-	int narows, nacols;
-	int MPIROOT=0;
-	char BLACS_LAYOUT='R', BLACS_SCOPE='A';
-	int ISRCPROC=0;
-	int desc[9];
-	
-	nFull=GlobalV::NLOCAL;
-	nblk=GlobalV::NLOCAL;	
-
-	// set mpi enviroment
-	int argc;
-	char **argv;
-//	MPI_Init(&argc, &argv);
-//    	MPI_Comm_size(MPI_COMM_WORLD, &nprocs);
-//    	MPI_Comm_rank(MPI_COMM_WORLD, &myid);
-	
-	mpierr=MPI_Bcast(&nFull, 1, MPI_INT, MPIROOT, MPI_COMM_WORLD);
-	mpierr=MPI_Bcast(&nblk, 1, MPI_INT, MPIROOT, MPI_COMM_WORLD);
-
-	// set blacs parameters 
-	for(npcols=int(sqrt(double(GlobalV::NPROC))); npcols>=2; --npcols)
-    	{
-        	if(GlobalV::NPROC%npcols==0) break;
-    	}
-    	nprows=GlobalV::NPROC/npcols;
-	
-	std::cout << "22222" <<std::endl;
-
-//	Cblacs_get(MPI_COMM_WORLD, 0, &my_blacs_ctxt);
-//    	Cblacs_gridinit(&my_blacs_ctxt, &BLACS_LAYOUT, nprows, npcols);
-//	Cblacs_gridinfo(my_blacs_ctxt, &nprows, &npcols, &myprow, &mypcol);
-
-//	blacs_get_(MPI_COMM_WORLD, 0, &my_blacs_ctxt);
-
-//	blacs_gridinit_(&my_blacs_ctxt, &BLACS_LAYOUT, &nprows, &npcols);
-//	blacs_gridinfo_(&my_blacs_ctxt, &nprows, &npcols, &myprow, &mypcol);
-
-
-	narows=numroc_(&nFull, &nblk, &myprow, &ISRCPROC, &nprows);
-    	nacols=numroc_(&nFull, &nblk, &mypcol, &ISRCPROC, &npcols);
-    	descinit_(desc, &nFull, &nFull, &nblk, &nblk, &ISRCPROC, &ISRCPROC, &my_blacs_ctxt, &narows, &info);
-
-
-	//init main matrix
-
-//	int subMatrixSize=narows*nacols;
-
-	int subMatrixSize=nFull*nFull;
-	std::cout << "333333" << std::endl;
-
-	// start
-	int allinfo;
-    	char transa, transb, side, uplo, diag;
-    	int isrc=1, jsrc=1;
-    	double alpha, beta;
-    	int matrixCols=nacols;
-    	int lda;
-	int ipiv;
-	int lwork, iwork, liwork;
-
-
-	double *S, *work, *a, *b;
-	S=new double[subMatrixSize];
-	work=new double[subMatrixSize];
-
-	MPI_Barrier(MPI_COMM_WORLD);
-
-	std::cout << subMatrixSize << std::endl;
-
-	for(int i=0; i<subMatrixSize; ++i) S[i]= real(GlobalC::LM.Sloc2[i]);
-//	for(int i=0; i<subMatrixSize; ++i) std::cout << S[i] <<std::endl;
-
-	for(int i=0; i<subMatrixSize; ++i) c[i]= real(GlobalC::LM.Sloc2[i]);
-
-	int inc=1;
-	ModuleBase::GlobalFunc::DCOPY(&subMatrixSize, &c, &inc, S, &inc);
-
-
-	std::cout << "44444" << std::endl;
-
-//	pdgetrf(&nFull, &nFull, S, &isrc, &jsrc, desc, &ipiv, &info);
-
-	std::cout << "55555"	<<std::endl;
-
-	MPI_Allreduce(&info, &allinfo, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD);
-	
-	std::cout << "66666" <<std::endl;
-
-	//calculate the inverse
-//	pdgetri(&nFull, S, &isrc, &jsrc, desc, &ipiv, work, &lwork, &iwork, &liwork, &info);
-	
-	std::cout << "77777" << std::endl;
-
-//    	pdgemm_(&S, &Htmp, &nFull, &nFull, &nFull,
-//            	&alpha, work, &isrc, &jsrc, desc,
-//                    	q,    &isrc, &jsrc, desc,
-//            	&beta,  b,    &isrc, &jsrc, desc);
-
-	
-
-	MPI_Allreduce(&info, &allinfo, 1, MPI_INT, MPI_SUM, MPI_COMM_WORLD);
-
-	std::cout << "88888"	<< std::endl;
-
-	MPI_Barrier(MPI_COMM_WORLD);
-
-	std::cout << "99999" << std::endl;
-
-//	blacs_gridexit_(&my_blacs_ctxt);
-
-//    	MPI_Finalize();
-
-	std::cout <<"end!!!!!"<< std::endl;
-    	return 0;
-}
-
-int globalIndex(int localIndex, int nblk, int nprocs, int myproc)
-{
-    int iblock, gIndex;
-    iblock=localIndex/nblk;
-    gIndex=(iblock*nprocs+myproc)*nblk+localIndex%nblk;
-    return gIndex;
-}
-
-
-int localIndex(int globalIndex, int nblk, int nprocs, int& myproc)
-{
-    myproc=int((globalIndex%(nblk*nprocs))/nblk);
-    return int(globalIndex/(nblk*nprocs))*nblk+globalIndex%nblk;
-}
-
-*/
 
 int Evolve_LCAO_Matrix::using_ScaLAPACK_complex_3(const int &ik, complex<double>** c, ModuleBase::ComplexMatrix &wfc_2d)const
 {
