+++ source/src_pw/diago_cg.cpp
@@ -499,7 +499,7 @@ double Diago_CG::ddot_real
     double *pL,*pR;
     pL=(double *)psi_L;
     pR=(double *)psi_R;
-    double result=LapackConnector::dot(dim2,pL,1,pR,1);
+    double result=BlasConnector::dot(dim2,pL,1,pR,1);
     if(reduce)  Parallel_Reduce::reduce_double_pool( result );
     return result;
     //======================================================================
